+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   * Added "auxiliary sources" option to support IDL sources that are loaded into the model but are
     not generated from.
 
-## 5.8.4
+## 5.8.5
 Release date: 2019-12-03
 ### Bug fixes:
   * Fixed Swift compilation issue for lambdas nested inside interfaces.